GEODE-6092: Simplify SingleHopClientExecutorLoggingIntegrationTest
