+++ geode-core/src/integrationTest/java/org/apache/geode/cache/client/internal/SingleHopClientExecutorLoggingIntegrationTest.java
@@ -1,0 +1,47 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more contributor license
+ * agreements. See the NOTICE file distributed with this work for additional information regarding
+ * copyright ownership. The ASF licenses this file to You under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the License. You may obtain a
+ * copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package org.apache.geode.cache.client.internal;
+
+import static org.apache.geode.test.awaitility.GeodeAwaitility.await;
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.contrib.java.lang.system.SystemOutRule;
+import org.junit.experimental.categories.Category;
+
+import org.apache.geode.test.junit.categories.ClientServerTest;
+import org.apache.geode.test.junit.categories.LoggingTest;
+
+/**
+ * Verifies that {@link SingleHopClientExecutor#submitTask} logs any exceptions thrown by task.
+ */
+@Category({ClientServerTest.class, LoggingTest.class})
+public class SingleHopClientExecutorLoggingIntegrationTest {
+
+  @Rule
+  public SystemOutRule systemOutRule = new SystemOutRule().enableLog();
+
+  @Test
+  public void submittedTaskShouldLogFailure() {
+    String message = "I am expecting this to be logged";
+
+    SingleHopClientExecutor.submitTask(() -> {
+      throw new RuntimeException(message);
+    });
+
+    await().untilAsserted(() -> assertThat(systemOutRule.getLog()).contains(message));
+  }
+}
